requirements.txt
click>=6.6
networkx>=1.11
nlpcube>=0.1
numpy>=1.11.2
editdistance==0.4